pyproject.toml
[tool.poetry]
name = "nataraz-studio"
version = "0.1.0"
description = "Nataraz Studio Website backend"
authors = ["Qiyas <qiyasturk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.15.2"
django = "^5.1.2"
pillow = "^11.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
